refactor : 기존 디렉터리 형식에 맞춰 클래스 파일 위치 변경
