--- src/main/java/com/jungle/navigation/chat/domain/RoomType.java
@@ -1,7 +1,0 @@
-package com.jungle.navigation.chat.domain;
-
-public enum RoomType {
-	DIRECT,
-	PRIVATE,
-	PUBLIC
-}
