+++ src/main/java/com/jungle/navigation/chat/persistence/entity/ChatRoom.java
@@ -1,7 +1,9 @@
-package com.jungle.navigation.chat.domain;
+package com.jungle.navigation.chat.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -19,18 +21,20 @@ import lombok.experimental.SuperBuilder;
 @ToString
 @SuperBuilder(toBuilder = true)
 @Entity
-@Table(name = RoomMember.PREFIX)
-public class RoomMember {
-	public static final String PREFIX = "room_member";
+@Table(name = ChatRoom.PREFIX)
+public class ChatRoom {
+	public static final String PREFIX = "room";
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = RoomMember.PREFIX + "_id")
+	@Column(name = ChatRoom.PREFIX + "_id")
 	private Long id;
 
-	@Column(name = "user_id")
-	private Long memberId;
+	@Enumerated(EnumType.STRING)
+	@Column(name = "room_type", nullable = false)
+	private RoomType roomType;
 
-	@Column(name = "room_id")
-	private Long chatRoomId;
+	public static ChatRoom of(RoomType roomType) {
+		return ChatRoom.builder().roomType(roomType).build();
+	}
 }
