+++ src/main/java/com/jungle/navigation/chat/persistence/entity/RoomType.java
@@ -1,0 +1,7 @@
+package com.jungle.navigation.chat.persistence.entity;
+
+public enum RoomType {
+	DIRECT,
+	PRIVATE,
+	PUBLIC
+}
